requirements.txt
transformers>=4.35.0
torch>=2.0.0
datasets>=2.14.0
accelerate>=0.23.0
huggingface-hub>=0.19.0
pyyaml>=6.0.0 